Added a warning popup based of the wireframes made the code efficient by adding components. added topic list overview tab Added styling for the signout button and fixed few layout issues(based of the wireframes) and other warnings that were showing up on the console
