+++ pages/admin-dashboard.js
@@ -3,11 +3,12 @@ import AdminOverview from "/components/AdminOverview";
 import styles from "../styles/Home.module.css";
 import Link from "next/link";
 import useSWR from "swr";
-import { signOut, useSession, getSession } from "next-auth/react";
+import { signOut, useSession } from "next-auth/react";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 import Button from "react-bootstrap/Button";
 import { Loading } from "../components/Loading";
+import WarningPopup from "../components/WarningPopup";
 
 function AdminDashboard() {
   const { data: session } = useSession();
@@ -35,10 +36,7 @@ function AdminDashboard() {
         <Loading />
       </div>
     );
-
   if (data["result"].length == 0) return <div>Not found</div>;
-  console.log("data");
-  console.log(data);
   const tutorialId = data["result"][0]["tutorial"];
 
   return (
@@ -46,7 +44,7 @@ function AdminDashboard() {
       <Head>
         <title>Staff Dashboard</title>
       </Head>
-      <h1 className={styles.heading}>Staff Dashboard</h1>
+      <h1 className={styles.heading}>Tutorial {tutorialId ? tutorialId[tutorialId?.length - 1] : null}</h1>
 
       <div
         style={{
@@ -56,16 +54,12 @@ function AdminDashboard() {
           marginTop: "-50px",
         }}
       >
-        {session ? session.user.name : ""} {tutorialId}{" "}
-        <Button onClick={() => signOut()}>
+        {session ? session.user.name : ""}
+        <Button onClick={() => signOut()} variant="danger" style={{ backgroundColor: "#FF595E", width: "250px" }}>
           <>Sign out</>
         </Button>{" "}
-        <Link href={`/create-topic-preferences?tutorialId=${tutorialId}`}>
-          <Button>
-            <>Create Topics</>
-          </Button>
-        </Link>
       </div>
+      <WarningPopup tutorialId={tutorialId} />
 
       <AdminOverview tutorialId={tutorialId} />
     </div>
